Release/2.3

+++ src/BlazorWasmAntivirusProtection.Tasks/RenameDlls.cs
@@ -1,6 +1,5 @@
 namespace BlazorWasmAntivirusProtection.Tasks
 {
-    using Brotli;
     using Microsoft.Build.Framework;
     using Microsoft.Build.Utilities;
     using System;
@@ -17,6 +16,8 @@ namespace BlazorWasmAntivirusProtection.Tasks
     {
         [Required]
         public string PublishDir { get; set; }
+        [Required]
+        public string BrotliCompressToolPath { get; set; }
 
         public string RenameDllsTo { get; set; } = "bin";
         public bool DisableRenamingDlls { get; set; }
@@ -81,12 +82,18 @@ namespace BlazorWasmAntivirusProtection.Tasks
                 if (File.Exists(bootJsonGzPath) && BlazorEnableCompression)
                 {
                     Log.LogMessage(MessageImportance.High, $"BlazorWasmAntivirusProtection: Recompressing \"{bootJsonGzPath}\"");
-                    GZipCompress(bootJsonPath, bootJsonGzPath);
+                    if(!GZipCompress(bootJsonPath, bootJsonGzPath))
+                    {
+                        return false;
+                    }
                 }
                 if (File.Exists(bootJsonBrPath) && BlazorEnableCompression)
                 {
                     Log.LogMessage(MessageImportance.High, $"BlazorWasmAntivirusProtection: Recompressing \"{bootJsonBrPath}\"");
-                    BrotliCompress(bootJsonPath, bootJsonBrPath);
+                    if(!BrotliCompress(bootJsonPath, bootJsonBrPath))
+                    {
+                        return false;
+                    }
                 }
             }
             
@@ -95,7 +102,7 @@ namespace BlazorWasmAntivirusProtection.Tasks
             return true;
         }
 
-        private void GZipCompress(string bootJsonPath, string bootJsonGzPath)
+        private bool GZipCompress(string bootJsonPath, string bootJsonGzPath)
         {
             try
             {
@@ -107,29 +114,45 @@ namespace BlazorWasmAntivirusProtection.Tasks
             }
             catch (Exception ex)
             {
-                Log.LogErrorFromException(ex);
+                if (File.Exists(bootJsonGzPath))
+                {
+                    File.Delete(bootJsonGzPath);
+                }
+                Log.LogErrorFromException(ex, true, true, null);
+                return false;
             }
+            return true;
         }
 
-        private void BrotliCompress(string bootJsonPath, string bootJsonBrPath)
+        private bool BrotliCompress(string bootJsonPath, string bootJsonBrPath)
         {
             try
             {
-                File.Delete(bootJsonBrPath);
-                var compressionLevel = Enum.TryParse<CompressionLevel>(CompressionLevel, out var level) ? level : System.IO.Compression.CompressionLevel.Optimal;
-                using var fileStream = File.OpenRead(bootJsonPath);
-                using var stream = File.Create(bootJsonBrPath);
-                fileStream.CompressToBrotli(stream, compressionLevel switch
+                // NOTE: This MSBuild Custom Task will run not only on .NET 6 or later but also on .NET Framework 4.x.
+                //       Therefore the `BrotliStream` class is not usable in this MSBuild Custom Task due to
+                //       the `BrotliStream` class is not provided on.NET Framework.Instead, we can do that
+                //       with execution as an out process.
+                var startInfo = new ProcessStartInfo
                 {
-                    System.IO.Compression.CompressionLevel.Optimal => 11,
-                    System.IO.Compression.CompressionLevel.Fastest => 5,
-                    System.IO.Compression.CompressionLevel.NoCompression=> 0,
-                });
+                    FileName = "dotnet",
+                    Arguments = $"exec \"{BrotliCompressToolPath}\" \"{bootJsonPath}\" \"{bootJsonBrPath}\" \"{CompressionLevel}\""
+                };
+                Log.LogMessage(MessageImportance.Low, $"{startInfo.FileName} {startInfo.Arguments}");
+                var process = Process.Start(startInfo);
+                process.WaitForExit();
+                if (process.ExitCode != 0) 
+                    throw new Exception($"The exit code of recompressing with Brotli command was not 0. (it was {process.ExitCode})");
             }
             catch (Exception ex)
             {
-                Log.LogErrorFromException(ex);
+                if (File.Exists(bootJsonBrPath))
+                {
+                    File.Delete(bootJsonBrPath);
+                }
+                Log.LogErrorFromException(ex, true, true, null);
+                return false;
             }
+            return true;
         }
 
         string ComputeSha256Hash(string rawData)
